+++ pkg/diskcache/database.go
@@ -3,7 +3,6 @@ package diskcache
 import (
 	"encoding/json"
 	"fmt"
-	"log"
 	"os"
 	"os/signal"
 	"syscall"